+++ include/linux/ioport.h
@@ -17,14 +17,14 @@
  * If you unload the driver, use release_region to free ports.
  */
 extern void reserve_setup(char *str, int *ints);
-extern int check_region(unsigned int from, unsigned int extent);
-extern void request_region(unsigned int from, unsigned int extent,const char *name);
-extern void release_region(unsigned int from, unsigned int extent);
+extern int check_region(unsigned long from, unsigned long extent);
+extern void request_region(unsigned long from, unsigned long extent,const char *name);
+extern void release_region(unsigned long from, unsigned long extent);
 extern int get_ioport_list(char *);
 
 #ifdef __sparc__
-extern unsigned int occupy_region(unsigned int base, unsigned int end,
-				  unsigned int num, unsigned int align,
+extern unsigned int occupy_region(unsigned long base, unsigned long end,
+				  unsigned long num, unsigned int align,
 				  const char *name);
 #endif
